feat(EST-3220-FE): move llm modals to shared
