+++ frontend/src/app/shared/components/index.ts
@@ -19,6 +19,7 @@ export * from './expand-panel/expand-panel.component';
 export * from './file-uploader/file-uploader.component';
 export * from './filters-list/filters-list.component';
 export * from './form-controls';
+export * from './llm-dialogs';
 //export * from './form-field-label/form-field-label.component';
 export * from './form-input/form-input.component';
 export * from './form-select/form-select.component';
