+++ frontend/src/app/shared/components/llm-dialogs/create-embedding-model-modal/create-embedding-model-modal.component.scss
