+++ frontend/src/app/shared/components/llm-dialogs/create-llm-model-modal/create-llm-model-modal.component.ts
@@ -4,27 +4,26 @@ import { ChangeDetectionStrategy, Component, DestroyRef, inject, OnInit, signal 
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { FormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
 import { MatTooltipModule } from '@angular/material/tooltip';
-import {
-    AppSvgIconComponent,
-    ButtonComponent,
-    CustomInputComponent,
-    TooltipComponent,
-    ValidationErrorsComponent,
-} from '@shared/components';
-import { LLMProvider, RealtimeModel } from '@shared/models';
-import { RealtimeModelsStorageService } from '@shared/services';
+import { LLMModel, LLMProvider } from '@shared/models';
+import { LlmModelsStorageService } from '@shared/services';
 import { getProviderIconPath } from '@shared/utils';
 import { finalize } from 'rxjs/operators';
 
 import { ToastService } from '../../../../services/notifications';
+import { AppSvgIconComponent } from '../../app-svg-icon/app-svg-icon.component';
+import { ValidationErrorsComponent } from '../../app-validation-errors/validation-errors.component';
+import { ButtonComponent } from '../../buttons';
+import { ToggleSwitchComponent } from '../../form-controls';
+import { CustomInputComponent } from '../../form-input/form-input.component';
+import { TooltipComponent } from '../../tooltip/tooltip.component';
 
-export interface CreateRealtimeModelDialogData {
+export interface CreateLlmModelDialogData {
     provider: LLMProvider;
-    model?: RealtimeModel;
+    model?: LLMModel;
 }
 
 @Component({
-    selector: 'app-create-realtime-model-modal',
+    selector: 'app-create-llm-model-modal',
     imports: [
         CommonModule,
         ReactiveFormsModule,
@@ -32,27 +31,32 @@ export interface CreateRealtimeModelDialogData {
         AppSvgIconComponent,
         CustomInputComponent,
         ButtonComponent,
+        ToggleSwitchComponent,
         TooltipComponent,
         ValidationErrorsComponent,
+        TooltipComponent,
     ],
-    templateUrl: './create-realtime-model-modal.component.html',
-    styleUrls: ['./create-realtime-model-modal.component.scss'],
+    templateUrl: './create-llm-model-modal.component.html',
+    styleUrls: ['./create-llm-model-modal.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class CreateRealtimeModelModalComponent implements OnInit {
+export class CreateLlmModelModalComponent implements OnInit {
     private dialogRef = inject(DialogRef);
-    private dialogData = inject<CreateRealtimeModelDialogData>(DIALOG_DATA);
+    private dialogData = inject<CreateLlmModelDialogData>(DIALOG_DATA);
     private destroyRef = inject(DestroyRef);
     private fb = inject(FormBuilder);
-    private realtimeModelsService = inject(RealtimeModelsStorageService);
+    private modelsStorageService = inject(LlmModelsStorageService);
     private toastService = inject(ToastService);
 
     isEditMode = !!this.dialogData.model;
-
     isSubmitting = signal(false);
 
     form = this.fb.group({
         name: [this.dialogData.model?.name ?? '', Validators.required],
+        baseUrl: [this.dialogData.model?.base_url ?? '', Validators.pattern(/^$|^https?:\/\/.+/i)],
+        deploymentId: [this.dialogData.model?.deployment_id ?? ''],
+        apiVersion: [this.dialogData.model?.api_version ?? ''],
+        isVisible: [this.dialogData.model?.is_visible ?? true],
     });
 
     provider = this.dialogData.provider;
@@ -80,13 +84,27 @@ export class CreateRealtimeModelModalComponent implements OnInit {
         const value = this.form.getRawValue();
         this.isSubmitting.set(true);
 
-        const name = (value.name || '').trim();
         const request$ = this.isEditMode
-            ? this.realtimeModelsService.patchModel(this.dialogData.model!.id, { name })
-            : this.realtimeModelsService.createModel({ name, provider: this.provider.id, is_custom: true });
+            ? this.modelsStorageService.patchModel(this.dialogData.model!.id, {
+                  name: (value.name || '').trim(),
+                  base_url: value.baseUrl?.trim() || null,
+                  deployment_id: value.deploymentId?.trim() || null,
+                  api_version: value.apiVersion?.trim() || null,
+                  is_visible: !!value.isVisible,
+              })
+            : this.modelsStorageService.createModel({
+                  name: (value.name || '').trim(),
+                  base_url: value.baseUrl?.trim() || null,
+                  deployment_id: value.deploymentId?.trim() || null,
+                  api_version: value.apiVersion?.trim() || null,
+                  llm_provider: this.provider.id,
+                  is_visible: !!value.isVisible,
+                  is_custom: true,
+                  predefined: false,
+              });
 
         request$.pipe(finalize(() => this.isSubmitting.set(false))).subscribe({
-            next: (result) => {
+            next: (result: LLMModel) => {
                 this.dialogRef.close(result);
             },
             error: (error) => {
