+++ frontend/src/app/shared/components/llm-dialogs/create-transcription-model-modal/create-transcription-model-modal.component.scss
