+++ frontend/src/app/shared/components/llm-dialogs/create-transcription-model-modal/create-transcription-model-modal.component.ts
@@ -1,31 +1,28 @@
 import { DIALOG_DATA, DialogRef } from '@angular/cdk/dialog';
 import { CommonModule } from '@angular/common';
-import { ChangeDetectionStrategy, Component, DestroyRef, inject, OnInit, signal } from '@angular/core';
+import { ChangeDetectionStrategy, Component, DestroyRef, inject, signal } from '@angular/core';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { FormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
 import { MatTooltipModule } from '@angular/material/tooltip';
-import {
-    AppSvgIconComponent,
-    ButtonComponent,
-    CustomInputComponent,
-    ToggleSwitchComponent,
-    TooltipComponent,
-    ValidationErrorsComponent,
-} from '@shared/components';
-import { LLMModel, LLMProvider } from '@shared/models';
-import { LlmModelsStorageService } from '@shared/services';
+import { GetRealtimeTranscriptionModelRequest, LLMProvider } from '@shared/models';
+import { TranscriptionModelsStorageService } from '@shared/services';
 import { getProviderIconPath } from '@shared/utils';
 import { finalize } from 'rxjs/operators';
 
 import { ToastService } from '../../../../services/notifications';
+import { AppSvgIconComponent } from '../../app-svg-icon/app-svg-icon.component';
+import { ValidationErrorsComponent } from '../../app-validation-errors/validation-errors.component';
+import { ButtonComponent } from '../../buttons';
+import { CustomInputComponent } from '../../form-input/form-input.component';
+import { TooltipComponent } from '../../tooltip/tooltip.component';
 
-export interface CreateLlmModelDialogData {
+export interface CreateTranscriptionModelDialogData {
     provider: LLMProvider;
-    model?: LLMModel;
+    model?: GetRealtimeTranscriptionModelRequest;
 }
 
 @Component({
-    selector: 'app-create-llm-model-modal',
+    selector: 'app-create-transcription-model-modal',
     imports: [
         CommonModule,
         ReactiveFormsModule,
@@ -33,31 +30,27 @@ export interface CreateLlmModelDialogData {
         AppSvgIconComponent,
         CustomInputComponent,
         ButtonComponent,
-        ToggleSwitchComponent,
         TooltipComponent,
         ValidationErrorsComponent,
     ],
-    templateUrl: './create-llm-model-modal.component.html',
-    styleUrls: ['./create-llm-model-modal.component.scss'],
+    templateUrl: './create-transcription-model-modal.component.html',
+    styleUrls: ['./create-transcription-model-modal.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class CreateLlmModelModalComponent implements OnInit {
+export class CreateTranscriptionModelModalComponent {
     private dialogRef = inject(DialogRef);
-    private dialogData = inject<CreateLlmModelDialogData>(DIALOG_DATA);
+    private dialogData = inject<CreateTranscriptionModelDialogData>(DIALOG_DATA);
     private destroyRef = inject(DestroyRef);
     private fb = inject(FormBuilder);
-    private modelsStorageService = inject(LlmModelsStorageService);
+    private transcriptionModelsService = inject(TranscriptionModelsStorageService);
     private toastService = inject(ToastService);
 
     isEditMode = !!this.dialogData.model;
+
     isSubmitting = signal(false);
 
     form = this.fb.group({
         name: [this.dialogData.model?.name ?? '', Validators.required],
-        baseUrl: [this.dialogData.model?.base_url ?? '', Validators.pattern(/^$|^https?:\/\/.+/i)],
-        deploymentId: [this.dialogData.model?.deployment_id ?? ''],
-        apiVersion: [this.dialogData.model?.api_version ?? ''],
-        isVisible: [this.dialogData.model?.is_visible ?? true],
     });
 
     provider = this.dialogData.provider;
@@ -85,27 +78,13 @@ export class CreateLlmModelModalComponent implements OnInit {
         const value = this.form.getRawValue();
         this.isSubmitting.set(true);
 
+        const name = (value.name || '').trim();
         const request$ = this.isEditMode
-            ? this.modelsStorageService.patchModel(this.dialogData.model!.id, {
-                  name: (value.name || '').trim(),
-                  base_url: value.baseUrl?.trim() || null,
-                  deployment_id: value.deploymentId?.trim() || null,
-                  api_version: value.apiVersion?.trim() || null,
-                  is_visible: !!value.isVisible,
-              })
-            : this.modelsStorageService.createModel({
-                  name: (value.name || '').trim(),
-                  base_url: value.baseUrl?.trim() || null,
-                  deployment_id: value.deploymentId?.trim() || null,
-                  api_version: value.apiVersion?.trim() || null,
-                  llm_provider: this.provider.id,
-                  is_visible: !!value.isVisible,
-                  is_custom: true,
-                  predefined: false,
-              });
+            ? this.transcriptionModelsService.patchModel(this.dialogData.model!.id, { name })
+            : this.transcriptionModelsService.createModel({ name, provider: this.provider.id, is_custom: true });
 
         request$.pipe(finalize(() => this.isSubmitting.set(false))).subscribe({
-            next: (result: LLMModel) => {
+            next: (result) => {
                 this.dialogRef.close(result);
             },
             error: (error) => {
