+++ frontend/src/app/shared/components/llm-dialogs/embedding-model-config-dialog/embedding-model-config-dialog.component.scss
