+++ frontend/src/app/shared/components/llm-dialogs/llm-model-config-dialog/llm-model-config-dialog.component.ts
@@ -6,14 +6,14 @@ import { LLMModel, LLMProvider, ModelTypes } from '@shared/models';
 import { LlmConfigStorageService } from '@shared/services';
 
 import { ToastService } from '../../../../services/notifications';
-import { InputNumberComponent } from '../../../../shared/components/app-input-number/input-number.component';
-import { ValidationErrorsComponent } from '../../../../shared/components/app-validation-errors/validation-errors.component';
-import { ButtonComponent } from '../../../../shared/components/buttons/button/button.component';
-import { IconButtonComponent } from '../../../../shared/components/buttons/icon-button/icon-button.component';
-import { CustomInputComponent } from '../../../../shared/components/form-input/form-input.component';
-import { JsonEditorFormFieldComponent } from '../../../../shared/components/json-editor/json-editor-form-field.component';
-import { KeyValueListComponent } from '../../../../shared/components/key-value-list/key-value-list.component';
-import { SliderWithStepperComponent } from '../../../../shared/components/slider-with-stepper/slider-with-stepper.component';
+import { InputNumberComponent } from '../../app-input-number/input-number.component';
+import { ValidationErrorsComponent } from '../../app-validation-errors/validation-errors.component';
+import { ButtonComponent } from '../../buttons';
+import { IconButtonComponent } from '../../buttons';
+import { CustomInputComponent } from '../../form-input/form-input.component';
+import { JsonEditorFormFieldComponent } from '../../json-editor/json-editor-form-field.component';
+import { KeyValueListComponent } from '../../key-value-list/key-value-list.component';
+import { SliderWithStepperComponent } from '../../slider-with-stepper/slider-with-stepper.component';
 import { LlmModelSelectorComponent } from '../llm-model-selector/llm-model-selector.component';
 
 interface DialogData {
