+++ frontend/src/app/shared/components/llm-dialogs/llm-model-selector/llm-model-selector.component.scss
