+++ frontend/src/app/shared/components/llm-dialogs/transcription-model-config-dialog/transcription-model-config-dialog.component.html
