+++ frontend/src/app/shared/components/llm-dialogs/voice-config-model/voice-model-config-dialog.component.scss
